backend/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.15.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
tensorflow-metal==1.2.0
termcolor==3.1.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
